+++ src/dreamscape/Discard.gd
@@ -1,14 +1,4 @@
-extends Pile
-
-# Declare member variables here. Examples:
-# var a = 2
-# var b = "text"
-
-
-# Called when the node enters the scene tree for the first time.
-func _ready():
-	pass # Replace with function body.
-
+extends DreamPile
 
 func reshuffle_in_deck():
 	var deck : Pile = cfc.NMAP.deck